requirements.txt
librespot-spotizerr
requests 